requirements.txt
pandas==1.5.2
plotly==5.11.0
python_highcharts==0.4.2
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.44
